+++ plone/app/contenttypes/tests/test_link.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 from datetime import datetime
+from plone.app.contenttypes.browser.widgets import LinkDataConverter
+from plone.app.contenttypes.browser.widgets import LinkWidget
 from plone.app.contenttypes.interfaces import ILink
 from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FUNCTIONAL_TESTING  # noqa
 from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_INTEGRATION_TESTING  # noqa
@@ -11,12 +13,14 @@ from plone.app.testing import TEST_USER_ID
 from plone.app.z3cform.interfaces import IPloneFormLayer
 from plone.dexterity.interfaces import IDexterityFTI
 from plone.testing.z2 import Browser
+from plone.uuid.interfaces import IUUID
 from Products.CMFCore.utils import getToolByName
 from zope.component import createObject
 from zope.component import getMultiAdapter
 from zope.component import queryUtility
 from zope.event import notify
 from zope.interface import alsoProvides
+from zope.schema import TextLine
 from zope.traversing.interfaces import BeforeTraverseEvent
 
 import unittest
@@ -270,3 +274,75 @@ class LinkFunctionalTest(unittest.TestCase):
         self.assertTrue(self.browser.url.endswith('my-special-link/view'))
         self.assertTrue('My link' in self.browser.contents)
         self.assertTrue('This is my link' in self.browser.contents)
+
+
+class LinkWidgetIntegrationTest(unittest.TestCase):
+
+    layer = PLONE_APP_CONTENTTYPES_INTEGRATION_TESTING
+
+    default_result = {'internal': u'',
+              'external': u'',
+              'email': u'',
+              'email_subject': u''}
+
+    def setUp(self):
+        self.portal = self.layer['portal']
+        self.request = self.layer['request']
+        self.request['ACTUAL_URL'] = self.portal.absolute_url()
+        self.response = self.request.response
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        lid = self.portal.invokeFactory('Link', 'link',
+                                        title='My Link',
+                                        description="This is my link.")
+        self.link = self.portal[lid]
+        self.link_field = TextLine()
+        self.widget = LinkWidget(self.request)
+
+    def test_dc_towidget_external(self):
+        url = 'http://www.example.org'
+        self.link.remoteUrl = url
+
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['external'] = url
+        self.assertEqual(converter.toWidgetValue(url), expected)
+
+    def test_dc_towidget_internal(self):
+        self.portal.invokeFactory('Document', 'doc1',
+                                  title='A document',
+                                  description="This is a document.")
+        doc1 = self.portal['doc1']
+        url = doc1.absolute_url()
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['internal'] = IUUID(doc1)
+        self.assertEqual(converter.toWidgetValue(url), expected)
+
+    def test_dc_towidget_mail(self):
+        url = u'mailto:foo@.example.org'
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['email'] = url[7:]   # mailto is cut
+        self.assertEqual(converter.toWidgetValue(url), expected)
+
+    def test_dc_towidget_mail_subject(self):
+        url = 'mailto:foo@.example.org?subject=A subject'
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['email'] = u'foo@.example.org'
+        expected['email_subject'] = u'A subject'
+        self.assertEqual(converter.toWidgetValue(url), expected)
+
+    def test_dc_illegal(self):
+        url = 'foo'
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['external'] = url
+        self.assertEqual(converter.toWidgetValue(url), expected)
+
+    def test_dc_var(self):
+        url = '${portal_url}/foo'
+        converter = LinkDataConverter(self.link_field, self.widget)
+        expected = self.default_result.copy()
+        expected['external'] = url
+        self.assertEqual(converter.toWidgetValue(url), expected)